Added Base Project, Record Timestamps, Updated KeyLogger log format...
Added support for recording the timestamp of the KeyStrokes
Created a Base KeyStroke project to house the common KeyPress class. Use this class when interacting with a Key Press.
Updated the HookKeylogger project to target NET 4.6 for the UnixTime methods.
Updated the HookKeylogger to use the new base KeyPress class to encode Key Presses in the Key Press Log.
Changed the ServerUpload program to just print the KeyLog created by the ServerUpload (Depreciated by Upcoming Code).
Added References to the Base project from most of the projects.
Added the Google.Protobuf package from NuGet.
Added Reference to the Google.Protobuf package from most of the projects.
Changed the name of the LoggerProgram file from Program to Logger.